requirements.txt
Flask==2.1.2
azure-cosmos==4.3.0
azure-monitor-opentelemetry-exporter==1.0.0b6
opentelemetry-instrumentation-flask==0.31b0
redis
